Bundle all the release steps together
