+++ .github/workflows/release.yml
@@ -39,22 +39,12 @@ jobs:
       - run: corepack enable
       - name: Install dependencies
         run: yarn install --immutable
-      - name: Get Next Version
-        id: next_version
-        run: |
-          export NEXT_VERSION="$(./scripts/getNextVersion.mjs)"
-          echo "Next Version: $NEXT_VERSION"
-          echo "version=$NEXT_VERSION" >> "$GITHUB_OUTPUT"
-        env:
-          GH_TOKEN: ${{ secrets.GH_TOKEN }}
       - name: Create Release
         run: yarn run release
         env:
           GH_TOKEN: ${{ secrets.GH_TOKEN }}
           NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
           SLACK_WEBHOOK_URL: ${{ secrets.SLACK_WEBHOOK_URL }}
-          NEXT_VERSION: ${{ steps.next_version.outputs.version }}
-          SENTRY_RELEASE: ${{ steps.next_version.outputs.version }}
           SENTRY_ORG: chromatic-lt
           SENTRY_PROJECT: cli
           SENTRY_AUTH_TOKEN: ${{ secrets.SENTRY_AUTH_TOKEN }}
