+++ package.json
@@ -73,11 +73,6 @@
     "build-storybook": "storybook build --stats-json",
     "build-test-storybook": "cross-env SMOKE_TEST=true storybook build -o test-storybook --stats-json",
     "build-subdir": "cd subdir ; yarn build ; cd ..",
-    "sentry": "yarn sentry:sourcemaps:cli && yarn sentry:sourcemaps:action && yarn sentry:release && yarn sentry:release:commits",
-    "sentry:sourcemaps:cli": "sentry-cli sourcemaps inject dist && sentry-cli sourcemaps upload dist",
-    "sentry:sourcemaps:action": "sentry-cli sourcemaps inject action && sentry-cli sourcemaps upload action",
-    "sentry:release": "sentry-cli releases new -p cli $NEXT_VERSION",
-    "sentry:release:commits": "sentry-cli releases set-commits --auto $NEXT_VERSION",
     "chromatic": "./dist/bin.js",
     "chromatic-prebuilt": "./dist/bin.js --storybook-build-dir=\"storybook-static\"",
     "chromatic-staging": "CHROMATIC_INDEX_URL=https://www.staging-chromatic.com ./dist/bin.js",
@@ -85,7 +80,7 @@
     "lint": "yarn lint:js bin-src node-src test-stories ./isChromatic.js ./isChromatic.mjs",
     "lint:js": "cross-env NODE_ENV=production eslint --cache --cache-location=.cache/eslint --report-unused-disable-directives",
     "lint:package": "sort-package-json",
-    "release": "./scripts/versionAndBuild.mjs && yarn sentry && yarn clean:sourcemaps && auto shipit",
+    "release": "./scripts/versionAndBuildForRelease.mjs && auto shipit",
     "prepack": "clean-package",
     "postpack": "clean-package restore",
     "publish-action": "./scripts/publish-action.mjs",
